backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.6
openai>=1.3.0
pdfplumber>=0.10.0
jinja2>=3.1.0
# weasyprint>=60.0  # Commented out - requires system dependencies, install separately if needed
cloudinary>=1.37.0
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0

